requirements.txt
openai
asyncpg
pytest
python-dotenv